fix scroll dispatcher circular dep
